Clean architecture for router
