+++ domain/model/pokemon.go
