+++ go.mod
@@ -5,7 +5,7 @@ go 1.16
 require (
 	model v0.0.0-00010101000000-000000000000 // indirect
 	modules v0.0.0-00010101000000-000000000000 // indirect
-	routes v0.0.0-00010101000000-000000000000
+	router v0.0.0-00010101000000-000000000000 // indirect
 	service/db v0.0.0-00010101000000-000000000000 // indirect
 	utils v0.0.0-00010101000000-000000000000
 )
@@ -16,6 +16,6 @@ replace utils => ./utils
 
 replace modules => ./modules
 
-replace routes => ./routes
-
 replace service/db => ./service/db
+
+replace router => ./router
