--- routes/go.mod
@@ -1,7 +1,0 @@
-module routes
-
-go 1.16
-
-require github.com/gorilla/mux v1.8.0
-
-replace modules => ../modules
